+++ Base/PLRenderer/src/Texture/TextureHandler.cpp
@@ -492,7 +492,7 @@ bool TextureHandler::Bind(uint32 nStage) const
 			FixedFunctions *pFixedFunctions = pTexture->GetTextureManager().GetRendererContext().GetRenderer().GetFixedFunctions();
 			if (pFixedFunctions) {
 				// Set texture matrix
-				pFixedFunctions->SetTransformState(static_cast<FixedFunctions::Transform::Enum>(FixedFunctions::Transform::Texture0 + nStage), m_pAnimationState->GetMatrix());
+				pFixedFunctions->SetTransformState(static_cast<FixedFunctions::Transform::Enum>(static_cast<uint32>(FixedFunctions::Transform::Texture0) + nStage), m_pAnimationState->GetMatrix());
 			}
 
 			// Done
@@ -506,7 +506,7 @@ bool TextureHandler::Bind(uint32 nStage) const
 			FixedFunctions *pFixedFunctions = pTexture->GetTextureManager().GetRendererContext().GetRenderer().GetFixedFunctions();
 			if (pFixedFunctions) {
 				// Set identity texture matrix
-				pFixedFunctions->SetTransformState(static_cast<FixedFunctions::Transform::Enum>(FixedFunctions::Transform::Texture0 + nStage), Matrix4x4::Identity);
+				pFixedFunctions->SetTransformState(static_cast<FixedFunctions::Transform::Enum>(static_cast<uint32>(FixedFunctions::Transform::Texture0) + nStage), Matrix4x4::Identity);
 			}
 
 			// Done
